apps/minerva/requirements.txt
# Minerva Backend Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
retell-sdk>=0.1.0
openai>=1.0.0
requests>=2.31.0
pydantic>=2.0.0

# 数据库
pymysql>=1.1.0
cryptography>=41.0.0







